Add visual guage meter

+++ app/src/main/java/com/hunting/fox/foxhunting/Activities/SettingsActivity.java
@@ -57,7 +57,7 @@ public class SettingsActivity extends AppCompatActivity {
         }
 
         try {
-            Settings.foxDuration = Float.parseFloat(edtFoxDuration.getText().toString());
+            Settings.foxDuration = Integer.parseInt(edtFoxDuration.getText().toString());
         } catch (Exception ex) {
             ex.printStackTrace();
         }
